requirements.txt
google-adk
google-cloud-spanner
uvicorn
fastapi
streamlit
requests
